+++ NEWS.md
@@ -1,5 +1,9 @@
 # testthat 0.11.0.9000
 
+* `find_reporter()` (and also all high-level testing functions) support a vector of reporters. For more than one reporter, a `MultiReporter` is created (#307, @krlmlr).
+
+* `with_reporter()` is used internally and gains new argument `start_end_reporter = TRUE` (@krlmlr, 355).
+
 * `expect_success()` and `expect_failure()` are new expectations designed
   specifically for testing other expectations (#368).